Fallthrough is no longer a regular rule.
